+++ sympy/rules/branch_strat_pure.py
